+++ cmd/data-node/commands/start/node.go
@@ -21,6 +21,8 @@ import (
 	"os/signal"
 	"syscall"
 
+	"code.vegaprotocol.io/vega/libs/subscribers"
+
 	"code.vegaprotocol.io/vega/datanode/admin"
 
 	embeddedpostgres "github.com/fergusstrange/embedded-postgres"
@@ -34,7 +36,6 @@ import (
 	"code.vegaprotocol.io/vega/datanode/networkhistory"
 	"code.vegaprotocol.io/vega/datanode/networkhistory/snapshot"
 	"code.vegaprotocol.io/vega/datanode/sqlstore"
-	"code.vegaprotocol.io/vega/datanode/subscribers"
 	"code.vegaprotocol.io/vega/libs/pprof"
 	"code.vegaprotocol.io/vega/logging"
 	"code.vegaprotocol.io/vega/paths"