+++ wallet/api/admin_send_raw_transaction_test.go
@@ -2,6 +2,7 @@ package api_test
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"testing"
 
@@ -113,6 +114,7 @@ func testAdminSendingRawTransactionWithValidParamsSucceeds(t *testing.T) {
 	sendingMode := "TYPE_SYNC"
 	network := newNetwork(t)
 	txHash := vgrand.RandomStr(64)
+	nodeHost := vgrand.RandomStr(5)
 
 	// setup
 	handler := newAdminSendRawTransactionHandler(t, func(hosts []string, retries uint64) (walletnode.Selector, error) {
@@ -120,6 +122,7 @@ func testAdminSendingRawTransactionWithValidParamsSucceeds(t *testing.T) {
 		nodeSelector := nodemocks.NewMockSelector(ctrl)
 		node := nodemocks.NewMockNode(ctrl)
 		nodeSelector.EXPECT().Node(ctx, gomock.Any()).Times(1).Return(node, nil)
+		node.EXPECT().Host().Times(1).Return(nodeHost)
 		node.EXPECT().SendTransaction(ctx, gomock.Any(), apipb.SubmitTransactionRequest_TYPE_SYNC).Times(1).Return(txHash, nil)
 		return nodeSelector, nil
 	})
@@ -294,7 +297,7 @@ func testAdminSendingRawTransactionWithFailedSendingFails(t *testing.T) {
 	})
 
 	// then
-	assertNetworkError(t, errorDetails, api.ErrTransactionFailed)
+	assertNetworkError(t, errorDetails, errors.New("the transaction failed: assert.AnError general error for testing"))
 	assert.Empty(t, result)
 }